New behavior
Changing a solvable value now turns off the solve flag.

Before, it would just temporarily turn it off during the solve, but it
seems more natural to keep it off.  Now, the user can just set values the
way they want and they don't have to worry about turning the solve flags off.
If they decide they want to solve the value again, it is easy enough to
turn the solving back on.