Working on changing how manipulators work.
Making it where manipulators are just separate objects in the scene.  Instead
of having manipulators attached to objects and having those objects
automatically move when the manipulator moves, it will be up to the caller
to check for manipulator movement and update the items in the scene
appropriately.  This should make it more natural to handle situations where
an object is being manipulated that can't move freely.  The manipulator will
stick with the mouse and the difference between where we're trying to put
the object and where it can actually go will be more clear.